auth/requirements.txt
Flask==2.3.2
redis
PyJWT
bcrypt
Flask-SQLAlchemy==3.1.1
psycopg2-binary
requests